fix: uint32 to uint24 for collectionIndexOf

+++ src/solc_0.8/assetERC1155/AssetBaseERC1155.sol
@@ -245,7 +245,7 @@ abstract contract AssetBaseERC1155 is WithSuperOperators, IERC1155 {
     /// @return the index/order at which the token `id` was minted in a collection.
     function collectionIndexOf(uint256 id) public view returns (uint256) {
         collectionOf(id); // this check if id and collection indeed was ever minted
-        return uint32((id & ERC1155ERC721Helper.NFT_INDEX) >> ERC1155ERC721Helper.NFT_INDEX_OFFSET);
+        return uint24((id & ERC1155ERC721Helper.NFT_INDEX) >> ERC1155ERC721Helper.NFT_INDEX_OFFSET);
     }
 
     /// end collection methods ---------------------------------------------------------------------------------------
